Nearly complete ArrayBag and driver
I finished making all of the ArrayBag methods and have tested all of them except for the toArray() which is giving me casting issues in the driver. I still need to learn what the purpose of the iterator is but that is the last thing I have to do before turning it in.